Refactor: move smaller components from Card into separate file
