+++ src/components/card/CardStatus.js
@@ -1,0 +1,51 @@
+import styled from 'styled-components';
+
+export function CardStatus({ status, species, type, className }) {
+  return (
+    <CardStatusContainer className={className}>
+      <StyledCardStatus status={status}>{status}</StyledCardStatus>
+      &nbsp;-&nbsp;
+      <CardSpecies>{species}</CardSpecies>
+      {type && <CardType>{type}</CardType>}
+    </CardStatusContainer>
+  );
+}
+
+const CardStatusContainer = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+`;
+
+const StyledCardStatus = styled.span`
+  display: flex;
+  align-items: center;
+  text-transform: capitalize;
+
+  &::before {
+    content: '';
+    display: block;
+    margin-right: 8px;
+    width: 9px;
+    height: 9px;
+    border-radius: 50%;
+    background-color: ${({ status }) => {
+      switch (status) {
+        case 'Alive':
+          return '#83bf46';
+        case 'Dead':
+          return '#ff5152';
+        default:
+          return '#968c9d';
+      }
+    }};
+  }
+`;
+
+const CardSpecies = styled.span``;
+
+const CardType = styled.p`
+  margin-top: 20px;
+  width: 100%;
+  color: #ddd;
+  font-size: 16px;
+`;
